fix(core): `selectLoading` returns `true` for not loading state
