+++ libs/core/state/src/operation/entity/public_api.ts
@@ -1,5 +1,5 @@
 export * from './adapter';
 export * from './selectors';
-export * from './helpers';
+export * from './helpers/create-fake-id';
 export { DaffOperationEntity } from './type';
 export { DaffOperationEntityState } from './state.type';
